Switch to Configuring In superset_config.py
The superset/config.py configuration file contains code that is
constantly being updated in newer versions. Replacing that file with
template in this repo might break the installation, if the template is
outdated.

Instead of replacing the superset/config.py file, copy over a
superset_config.py template that will override variables in the main
configuration file.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>