Set PYTHONPATH when configuring SUperset
This is because calling the Superset/Flask commands would otherwise
not use our custom config file.  This is bad because the commands would
not work as expected e.g. they are likely to be run against an SQLite
database (which is Superset's default) instead of the database defined
in our custom config file.

This essentially reverts 548535fa25947048c66021354042dda6b5bfe9ff